+++ packages/react-theme-default/src/components/SCCourses.tsx
@@ -105,9 +105,6 @@ const Component = {
             minWidth: '250px',
             [theme.breakpoints.down('md')]: {
               width: '100%'
-            },
-            '& .SCCourse-placeholder-root': {
-              height: '250px'
             }
           }
         }
@@ -121,17 +118,16 @@ const Component = {
     }),
     skeletonRoot: ({theme}: any) => ({
       '& .SCCourses-skeleton-courses': {
-        justifyContent: 'center',
         marginTop: theme.spacing(2),
         '& .SCCourses-skeleton-item': {
           paddingTop: theme.spacing(2)
         }
       }
     }),
-    coursesChipRoot: ({theme, showForMe}: any) => ({
+    coursesChipRoot: ({theme, showForMe, showMyCourses}: any) => ({
       height: theme.spacing(5.25),
       borderRadius: theme.spacing(0.5),
-      color: showForMe ? theme.palette.common.white : theme.palette.text.primary,
+      color: showForMe || showMyCourses ? theme.palette.common.white : theme.palette.text.primary,
       '& .MuiIcon-root': {
         fontSize: '1rem',
         color: theme.palette.common.white